Added real continuity check to layer 1 edge extraction. Various changes and improvements.

+++ 2D/CHOP/Prototype/utilities/calculatePooledPositions.m
@@ -1,5 +1,5 @@
 function [ positions ] = calculatePooledPositions( precisePositions, poolFactor, poolDim, stride )
-     tempPositions = floor((precisePositions-1) / stride);
+     tempPositions = floor((double(precisePositions)-1) / stride);
      tempPositions = floor(tempPositions / (poolDim^poolFactor));
-     positions = tempPositions + 1;
+     positions = int32(tempPositions + 1);
 end
